feat(generate): VAE decode 精度改为用户选项，默认 bf16
- settings.generate.vae_precision（bf16 / fp32），Settings 测试 tab 新增下拉；
  bf16 默认对齐 ComfyUI 在现代 GPU 上的 auto VAE dtype
- force_comfy_parity_runtime_config 不再强制 fp32：保留调用方传入值，
  缺省 bf16（DEFAULT_VAE_PRECISION）
- GenerateConfig 增加 vae_precision 字段（Literal）；enqueue 时从 settings 注入
- fp32 时仍走 decode 前 offload DiT/Qwen（上一 commit 已按 VAE 精度门控，
  bf16 默认路径零搬运开销）
- 测试：默认 bf16 / 显式 fp32 保留两组断言；pytest 2196 passed，
  tsc --noEmit 干净，vitest 366 passed

+++ tests/test_anima_daemon_comfy_parity_runtime.py
@@ -52,7 +52,7 @@ def test_daemon_model_cache_preserves_user_backend_and_sets_comfy_style_dtype(mo
 
     assert loads[0]["backend"] == "flash_attn"
     assert loads[0]["precision"] == "bf16"
-    assert loads[0]["vae_precision"] == "fp32"
+    assert loads[0]["vae_precision"] == "bf16"
     assert loads[0]["text_encoder_backend"] == "comfy_qwen3"
     assert loads[0]["t5_tokenizer_backend"] == "fast"
